Final adjustments on the Rust code - Separate the code into 3 modules   - lib.rs: the content related to the library itself   - projectors.rs: contains the code related to generating projections   - utils.rs: contains simple utilities - Small refinements to improve readability
